requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.1
python-dotenv==1.0.0
pydantic==2.5.3
asyncpg==0.29.0
psycopg2-binary==2.9.9
pytest
pytest-asyncio